1.29 环信对接完成

+++ app/src/main/java/com/example/admin/zgapplication/mvp/module/AgentDetailResponse.java
@@ -45,7 +45,7 @@ public class AgentDetailResponse extends BaseResponse<AgentDetailResponse.DataBe
         private String house_sum;
         private String visit_sum;
         private String order_sum;
-        private int score;
+        private float score;
         private String agent_count;
         private String house_count;
         private List<String> label;
@@ -148,11 +148,11 @@ public class AgentDetailResponse extends BaseResponse<AgentDetailResponse.DataBe
             this.order_sum = order_sum;
         }
 
-        public int getScore() {
+        public float getScore() {
             return score;
         }
 
-        public void setScore(int score) {
+        public void setScore(float score) {
             this.score = score;
         }
